déplacement des fonctions `parcours` dans un fichier à part de la class
