+++ Arbres/parcours.py
@@ -1,0 +1,65 @@
+from classe_noeud import *
+
+def parcours_infixe(self: Noeud):
+    """ affiche les valeurs de chaque noeud exploré séparées par des espaces """
+    if self == Noeud.arbre_vide:
+        return ""
+    return (
+        parcours_infixe(self.gauche()) +
+        str(self.valeur()) + " " +
+        parcours_infixe(self.droit())
+    )
+
+
+def parcours_prefixe(self: Noeud):
+    """ affiche les valeurs de chaque noeud exploré séparées par des espaces """
+    if self == Noeud.arbre_vide:
+        return ""
+    return (
+        str(self.valeur()) + " " +
+        parcours_prefixe(self.gauche()) +
+        parcours_prefixe(self.droit())
+    )
+
+
+def parcours_posfixe(self: Noeud):
+    if self == Noeud.arbre_vide:
+        return ""
+    return (
+        parcours_posfixe(self.gauche()) +
+        parcours_posfixe(self.droit()) +
+        str(self.valeur()) + " "
+    )
+
+
+A_g = Noeud("D", None, None)
+A_N = None
+A_1 = Noeud("A", Noeud("B", None, None), None)
+A_2 = Noeud("A", Noeud("B", None, None), Noeud("D", None, None))
+A_3 = Noeud("A", Noeud("B", None, Noeud("C", None, None)),
+            Noeud("D", None, None))
+A_4 = Noeud(1,
+            Noeud(2,
+                  Noeud(4,
+                        Noeud(8, Noeud.arbre_vide, Noeud.arbre_vide),
+                        Noeud(9, Noeud.arbre_vide, Noeud.arbre_vide)),
+                  Noeud(5,
+                        Noeud(10, Noeud.arbre_vide, Noeud.arbre_vide),
+                        Noeud.arbre_vide)),
+            Noeud(3,
+                  Noeud(6, Noeud.arbre_vide, Noeud.arbre_vide),
+                  Noeud(7, Noeud.arbre_vide, Noeud.arbre_vide)))
+
+print(A_3.gauche().droit().valeur())
+print(A_3.est_feuille())
+print(compte_feuilles(A_N))
+print(compte_feuilles(A_g))
+print(compte_feuilles(A_3))
+print(compte_feuilles(A_4))
+print(taille(A_N))
+print(hauteur(A_4))
+print(hauteur(A_2))
+print(hauteur(A_3))
+print(parcours_infixe(A_4))
+print(parcours_prefixe(A_4))
+print(parcours_posfixe(A_4))
